% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adk_metadata.R
\name{adk_metadata}
\alias{adk_metadata}
\title{Get data table metadata info}
\usage{
adk_metadata(data_name)
}
\arguments{
\item{data_name}{character name of dataset. See \code{\link{adk_data}} documentation for dataset names.}
}
\value{
Metadata about each dataset. Includes units and long-name of parameters
}
\examples{

#Get chemistry metadata
adk_metadata('chem')

}
