% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adk_lake_shapes.R
\name{adk_shape}
\alias{adk_shape}
\title{Return path to Adirondack Park Shapefile}
\usage{
adk_shape()
}
\description{
Returns the path to the shapefile for the Adirondack Park
outline (The "Blue Line"). Returns the path to a locally stored shapefile
that can be used for mapping and analysis.
}
\examples{
\dontrun{
#' library(sf)
bl = sf::read_sf(adklakedata::adk_shape())
lakes = sf::read_sf(adklakedata::adk_lake_shapes())
plot(st_geometry(bl))
plot(st_geometry(lakes), add=TRUE, col='blue')
}


}
