% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dl_data.R
\name{check_dl_data}
\alias{check_dl_data}
\title{Download lake data from internet}
\usage{
check_dl_data()
}
\description{
Check that we have local cache of ADK lake data.
If it is not locally available, download the data
from the internet and prepare it for local use. This only
needs to be run once for each install of the package. Note:
you will be required to re-download data when a new version of
the package is released. This ensures stale data are not being
accidentally used.
}
