% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dl_file.R
\name{check_dl_file}
\alias{check_dl_file}
\title{Verify and download data files}
\usage{
check_dl_file(master_file, fname = NULL, md5check = TRUE, dest = local_path())
}
\arguments{
\item{master_file}{Character path to master file}

\item{fname}{Character vector of specific file names to check}

\item{md5check}{boolean}

\item{dest}{Character path to download destination}
}
\description{
Checks if local data files as defined in master file exist and match MD5 hash. Downloads data if necessary.
}
