% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dl_file.R
\name{create_dl_file}
\alias{create_dl_file}
\title{Create master file for use in check_dl_file}
\usage{
create_dl_file(master_file, urllist, filelist)
}
\arguments{
\item{master_file}{Character path to export csv}

\item{urllist}{Character vector of URLs}

\item{filelist}{Character vector of files}
}
