% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata}
\usage{
get_metadata("chem")
}
\arguments{
\item{dataset_name}{character name of dataset. See \code{\link{adk_data}} documentation for dataset names.}
}
\value{
Metadata about dataset
}
