% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dl_data.R
\name{local_path}
\alias{local_path}
\title{Get local file path}
\usage{
local_path()
}
\value{
Path to local file cache location
}
\description{
Data files are locally cached (they are too large to be distributed
with the CRAN package). These cached files are stored in your user
data directory, or a custom directory set using \code{set_local_path}.
}
\examples{
# set custom path to local temp directory
set_local_path(tempdir())

#returns current local path directory
local_path()

}
