% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adk_data.R
\name{adk_data}
\alias{adk_data}
\title{Load ADK Data}
\usage{
adk_data(data_name)
}
\arguments{
\item{data_name}{A string choosing the data to load. Can be "chem", "rotifer", "phyto", "crustacean", "tempdo", or "meta".}
}
\description{
Loads data from locally downloaded CSV files. Run check_dl_data before using this function.
}
