% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adk_lake_shapes.R
\name{adk_lake_shapes}
\alias{adk_lake_shapes}
\title{Return path to Lake Polygons Shapefile}
\usage{
adk_lake_shapes()
}
\description{
Returns the path to the shapefile for the study Lake polygons.
The source is a locally stored shapefile
that can be used for mapping and analysis.
}
\examples{
library(sf)
bl = read_sf(adklakedata::adk_shape())
lakes = read_sf(adklakedata::adk_lake_shapes())
plot(st_geometry(bl))
plot(st_geometry(lakes), add=TRUE, col='blue')

}
