% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adk_lakes.R
\name{adk_lakes}
\alias{adk_lakes}
\title{List of lakes with attributes}
\usage{
adk_lakes()
}
\description{
Returns a data.frame of lake info. Includes common info like
lake location (lat/lon), lake name, and numerical site ID.
}
\examples{
\dontrun{
sites = adk_lakes()
}

}
