% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adk_lakes.R
\name{adk_lakes}
\alias{adk_lakes}
\title{Load the lake shapefile for ADK}
\usage{
adk_lakes()
}
\description{
This function loads the primary geospatial boundary data for the Adirondack parks.
}
\examples{
\dontrun{
sites = adk_lakes()
}

}
