% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dl_data.R
\name{adk_shape}
\alias{adk_shape}
\title{Return path to Adirondack Park Shapefile}
\usage{
adk_shape()
}
\description{
Downloads and unpacks the shapefile for the Adirondack Park
outline shapefile. Returns the path to a locally stored shapefile
that can be used for mapping an analysis.
}

